[MOD] : auth 파일 분리
