+++ app/src/main/java/org/sopt/dosopttemplate/data/entity/request/RequestLoginDto.kt
@@ -1,14 +1,12 @@
-package org.sopt.dosopttemplate.data.model.request
+package org.sopt.dosopttemplate.data.entity.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignUpDto(
+data class RequestLoginDto(
     @SerialName("username")
     val username: String,
-    @SerialName("nickname")
-    val nickname: String,
     @SerialName("password")
-    val password: String,
+    val password: String
 )
