+++ app/src/main/java/org/sopt/dosopttemplate/data/entity/response/ResponseFollowerDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.model.response
+package org.sopt.dosopttemplate.data.entity.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
