+++ app/src/main/java/org/sopt/dosopttemplate/data/entity/response/ResponseSignUpDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.model.response
+package org.sopt.dosopttemplate.data.entity.response
 
 data class ResponseSignUpDto(
     val success: Unit
