+++ app/src/main/java/org/sopt/dosopttemplate/data/entity/service/FollowerService.kt
@@ -1,0 +1,13 @@
+package org.sopt.dosopttemplate.data.entity.service
+
+import org.sopt.dosopttemplate.data.entity.response.ResponseFollowerDto
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface FollowerService {
+    @GET("api/users")
+    suspend fun follower(
+        @Query("page") num: Int = 2
+    ): Call<ResponseFollowerDto>
+}
