+++ app/src/main/java/org/sopt/dosopttemplate/data/entity/service/SignUpService.kt
@@ -1,0 +1,12 @@
+package org.sopt.dosopttemplate.data.entity.service
+
+import org.sopt.dosopttemplate.data.entity.request.RequestSignUpDto
+import retrofit2.http.Body
+import retrofit2.http.POST
+
+interface SignUpService {
+    @POST("api/v1/members")
+    suspend fun signUp(
+        @Body signUpReq: RequestSignUpDto,
+    )
+}
