--- app/src/main/java/org/sopt/dosopttemplate/data/model/service/AuthService.kt
@@ -1,20 +1,0 @@
-package org.sopt.dosopttemplate.data.model.service
-
-import org.sopt.dosopttemplate.data.model.request.RequestLoginDto
-import org.sopt.dosopttemplate.data.model.request.RequestSignUpDto
-import org.sopt.dosopttemplate.data.model.response.ResponseLoginDto
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface AuthService {
-    @POST("api/v1/members")
-    suspend fun signUp(
-        @Body signUpReq: RequestSignUpDto,
-    )
-
-    @POST("api/v1/members/sign-in")
-    suspend fun login(
-        @Body request: RequestLoginDto,
-    ): Response<ResponseLoginDto>
-}
